Potential fix for code scanning alert no. 2: Implicit narrowing conversion in compound assignment

+++ src/shutterencoder/functions/Merge.java
@@ -66,7 +66,7 @@ public class Merge extends Shutter {
 
 					try {
 						
-						int totalLength = 0;
+						long totalLength = 0L;
 						frame.setCursor(Cursor.getPredefinedCursor(Cursor.WAIT_CURSOR));			
 						PrintWriter writer = new PrintWriter(listeBAB, "UTF-8");
